Change build dir to build/platform/architecture
Raft can be used to cross-compile. If a tool is consuming artifacts
built by Raft, it needs to know where each build is located. One example
is using gradle to pull all of the libraries into the APK. Nesting the
build directories allows tools to know the location by convention.